% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Title}
\description{
Title
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{X}}{matrix.}

\item{\code{Regressions_coefficients}}{vector.}

\item{\code{The_fitted_values}}{matrix.}

\item{\code{The_residuals}}{matrix.}

\item{\code{The_degrees_of_freedom}}{numeric.}

\item{\code{The_residual_variance}}{numeric.}

\item{\code{The_variance_of_the_regression_coefficients}}{matrix.}

\item{\code{The_t_values}}{matrix.}
}}


